pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-engineering-realtime-pipeline"
version = "0.1.0"
description = "Real-time data engineering pipeline for MTA GTFS feeds and NYC taxi data analysis"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Data Engineer", email = "engineer@example.com" }]

dependencies = [
    # Core data processing
    "confluent-kafka>=2.3.0",
    "protobuf>=4.25.1",
    "pandas>=2.1.0",
    "requests>=2.31.0",
    
    # Database and storage
    "duckdb>=0.9.2",
    
    # GTFS-realtime specific
    "gtfs-realtime-bindings>=1.0.0",
    
    # For visualization
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",
    
    # Jupyter integration
    "jupyter>=1.1.1",
    "notebook>=7.4",
    "jupyterlab>=4.0.0",
    
    # Optional: cloud integration
    "boto3>=1.34.0",      # For AWS
    "google-cloud-storage>=2.13.0",  # For GCP
    "azure-storage-blob>=12.18.0",   # For Azure
    
    # Environment utilities
    "python-dotenv>=1.0.0",

    # Stream processing and query engines
    "trino>=0.325.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.3.0",
    "flake8>=7.0.0",
    "mypy>=1.5.1",
]

orchestration = [
    "apache-airflow>=2.7.0",
    "prefect>=2.13.0",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false